backend/requirements.txt
# https://www.python.org/dev/peps/pep-0440/#compatible-release
Django~=3.1
psycopg2~=2.8

gunicorn~=19.9
uvicorn~=0.12.2
django-rest-framework~=0.1.0
djangorestframework~=3.11.0
djangorestframework-jwt==1.11.0
drf-yasg~=1.17.1
channels~=2.4.0
channels-redis~=3.2.0
django-cors-headers==3.5.0
